:recycle: refactor: Migrate Redux slices to use use cases
